+++ ChangeLog.md
@@ -3,6 +3,7 @@
 ## Unreleased changes
 
  - Disable sessions since we don't use them.
+ - Add query parameter `includeRawRejectReason` to the `accTransactions` query.
 
 ## 0.5.0
  - Make the GTU drop functionality optional.